resolve a bunch of circular dependencies
