+++ shared-data/js/helpers/getLabwareHasQuirk.ts
@@ -1,0 +1,9 @@
+import { LabwareDefinition } from '../types'
+
+export function getLabwareHasQuirk(
+  labwareDef: LabwareDefinition,
+  quirk: string
+): boolean {
+  const quirks = labwareDef.parameters.quirks
+  return quirks ? quirks.includes(quirk) : false
+}
