+++ shared-data/js/helpers/getWellNamePerMultiTip.ts
@@ -1,7 +1,9 @@
 import range from 'lodash/range'
 
+import { getLabwareHasQuirk } from '././getLabwareHasQuirk'
 import { get96Channel384WellPlateWells } from './get96Channel384WellPlateWells'
-import { getLabwareHasQuirk, orderWells, sortWells } from './index'
+import { orderWells } from './orderWells'
+import { sortWells } from './sortWells'
 
 import type { LabwareDefinition } from '../types'
